+++ platform/core/src/DICOMSR/parseDicomStructuredReport.js
@@ -1,6 +1,10 @@
 import dcmjs from 'dcmjs';
+import classes from '../classes';
+import parseSCOORD3D from './SCOORD3D/parseSCOORD3D';
 
 import findInstanceMetadataBySopInstanceUID from './utils/findInstanceMetadataBySopInstanceUid';
+
+const { LogManager } = classes;
 
 /**
  * Function to parse the part10 array buffer that comes from a DICOM Structured report into measurementData
@@ -9,20 +13,40 @@ import findInstanceMetadataBySopInstanceUID from './utils/findInstanceMetadataBy
  *
  * @param {ArrayBuffer} part10SRArrayBuffer
  * @param {Array} displaySets
+ * @param {object} external
  * @returns
  */
-const parseDicomStructuredReport = (part10SRArrayBuffer, displaySets) => {
-  // Get the dicom data as an Object
+const parseDicomStructuredReport = (
+  part10SRArrayBuffer,
+  displaySets,
+  external
+) => {
+  if (external && external.servicesManager) {
+    parseSCOORD3D({ servicesManager: external.servicesManager, displaySets });
+  }
 
+  // Get the dicom data as an Object
   const dicomData = dcmjs.data.DicomMessage.readFile(part10SRArrayBuffer);
   const dataset = dcmjs.data.DicomMetaDictionary.naturalizeDataset(
     dicomData.dict
   );
 
   const { MeasurementReport } = dcmjs.adapters.Cornerstone;
-  const storedMeasurementByToolType = MeasurementReport.generateToolState(
-    dataset
-  );
+
+  let storedMeasurementByToolType;
+  try {
+    storedMeasurementByToolType = MeasurementReport.generateToolState(dataset);
+  } catch (error) {
+    const seriesDescription = dataset.SeriesDescription || '';
+    LogManager.publish(LogManager.EVENTS.OnLog, {
+      title: `Failed to parse ${seriesDescription} measurement report`,
+      type: 'warning',
+      message: error.message || '',
+      notify: true,
+    });
+    return;
+  }
+
   const measurementData = {};
   let measurementNumber = 0;