+++ platform/core/src/redux/actions.test.js
@@ -11,8 +11,6 @@ describe('actions', () => {
       'setLayout',
       'clearViewportSpecificData',
       'setActiveViewportSpecificData',
-      'setStudyLoadingProgress',
-      'clearStudyLoadingProgress',
       'setUserPreferences',
       'setExtensionData',
       'setTimepoints',