+++ platform/core/src/utils/loadAndCacheDerivedDisplaySets.js
@@ -49,7 +49,12 @@ import studyMetadataManager from './studyMetadataManager';
  * @param {object} snackbar
  * @returns void
  */
-async function loadAndCacheDerivedDisplaySets(referencedDisplaySet, studies, logger, snackbar) {
+async function loadAndCacheDerivedDisplaySets(
+  referencedDisplaySet,
+  studies,
+  logger,
+  snackbar
+) {
   const { StudyInstanceUID, SeriesInstanceUID } = referencedDisplaySet;
   const promises = [];
   const studyMetadata = studyMetadataManager.get(StudyInstanceUID);
@@ -80,60 +85,111 @@ async function loadAndCacheDerivedDisplaySets(referencedDisplaySet, studies, log
   });
 
   // For each type, see if any are loaded, if not load the most recent.
-  await Promise.all(Object.keys(displaySetsPerModality).map(async (key) => {
-    const displaySets = displaySetsPerModality[key];
+  await Promise.all(
+    Object.keys(displaySetsPerModality).map(async key => {
+      const displaySets = displaySetsPerModality[key];
 
-    const isLoaded = displaySets.some(displaySet => displaySet.isLoaded);
-    if (isLoaded) {
-      return;
-    }
-
-    if (displaySets.some(displaySet => displaySet.loadError)) {
-      return;
-    }
-
-    // find most recent and load it.
-    let recentDateTime = 0;
-    let recentDisplaySet = displaySets[0];
-
-    displaySets.forEach(displaySet => {
-      const dateTime = Number(
-        `${displaySet.SeriesDate}${displaySet.SeriesTime}`
-      );
-      if (dateTime > recentDateTime) {
-        recentDateTime = dateTime;
-        recentDisplaySet = displaySet;
+      const isLoaded = displaySets.some(displaySet => displaySet.isLoaded);
+      if (isLoaded) {
+        return;
       }
-    });
 
-    try {
-      if (recentDisplaySet.hasOwnProperty('getSourceDisplaySet') &&
-        typeof recentDisplaySet.getSourceDisplaySet === 'function') {
-        await recentDisplaySet.getSourceDisplaySet(studies);
-      } else {
-        await recentDisplaySet.load(referencedDisplaySet, studies);
+      if (displaySets.some(displaySet => displaySet.loadError)) {
+        return;
       }
-    } catch (error) {
-      recentDisplaySet.isLoaded = false;
-      recentDisplaySet.loadError = true;
-      logger.error({ error, message: error.message });
-      snackbar.show({
-        title: 'Error loading derived display set:',
-        message: error.message,
-        type: 'error',
-        error,
-        autoClose: false,
+
+      // find most recent and load it.
+      let recentDateTime = 0;
+      let recentDisplaySet = displaySets[0];
+
+      displaySets.forEach(displaySet => {
+        const dateTime = Number(
+          `${displaySet.SeriesDate}${displaySet.SeriesTime}`
+        );
+        if (dateTime > recentDateTime) {
+          recentDateTime = dateTime;
+          recentDisplaySet = displaySet;
+        }
       });
-    }
-  }));
+
+      try {
+        if (
+          recentDisplaySet.hasOwnProperty('getSourceDisplaySet') &&
+          typeof recentDisplaySet.getSourceDisplaySet === 'function'
+        ) {
+          if (recentDisplaySet.Modality === 'SEG' && logger) {
+            const onDisplaySetLoadFailureHandler = error => {
+              logger.error({ error, message: error.message });
+              snackbar.show({
+                title: 'DICOM Segmentation Loader',
+                message: error.message,
+                type: 'error',
+                autoClose: true,
+              });
+            };
+
+            let activatedLabelmapIndex = -1;
+            while (activatedLabelmapIndex == -1) {
+              const {
+                referencedDisplaySet,
+                activatedLabelmapPromise,
+              } = await recentDisplaySet.getSourceDisplaySet(
+                studies,
+                true,
+                onDisplaySetLoadFailureHandler
+              );
+
+              activatedLabelmapIndex = await activatedLabelmapPromise;
+              const selectionFired = new CustomEvent(
+                'extensiondicomsegmentationsegselected',
+                {
+                  detail: { activatedLabelmapIndex: activatedLabelmapIndex },
+                }
+              );
+              document.dispatchEvent(selectionFired);
+
+              const lastDateTime = Number(
+                `${recentDisplaySet.SeriesDate}${recentDisplaySet.SeriesTime}`
+              );
+              recentDateTime = 0;
+              displaySets.forEach(displaySet => {
+                const dateTime = Number(
+                  `${displaySet.SeriesDate}${displaySet.SeriesTime}`
+                );
+                if (dateTime > recentDateTime && dateTime < lastDateTime) {
+                  recentDateTime = dateTime;
+                  recentDisplaySet = displaySet;
+                }
+              });
+            }
+          } else {
+            await recentDisplaySet.getSourceDisplaySet(studies);
+          }
+        } else {
+          await recentDisplaySet.load(referencedDisplaySet, studies);
+        }
+      } catch (error) {
+        recentDisplaySet.isLoaded = false;
+        recentDisplaySet.loadError = true;
+        logger.error({ error, message: error.message });
+        snackbar.show({
+          title: 'Error loading derived display set:',
+          message: error.message,
+          type: 'error',
+          error,
+          autoClose: false,
+        });
+      }
+    })
+  );
 
   /*
-  * TODO: Improve the way we notify parts of the app
-  * that depends on derived display sets to be loaded.
-  * (Implement pubsub for better tracking of derived display sets)
-  */
+   * TODO: Improve the way we notify parts of the app
+   * that depends on derived display sets to be loaded.
+   * (Implement pubsub for better tracking of derived display sets)
+   */
   const event = new CustomEvent('deriveddisplaysetsloadedandcached');
   document.dispatchEvent(event);
-};
+}
 
 export default loadAndCacheDerivedDisplaySets;