+++ platform/viewer/public/config/default.js
@@ -5,6 +5,16 @@ window.config = {
   showStudyList: true,
   filterQueryParam: false,
   disableServersCache: false,
+  studyPrefetcher: {
+    enabled: true,
+    order: 'closest',
+    displaySetCount: 3,
+    preventCache: false,
+    prefetchDisplaySetsTimeout: 300,
+    maxNumPrefetchRequests: 100,
+    displayProgress: true,
+    includeActiveDisplaySet: true,
+  },
   servers: {
     dicomWeb: [
       {
@@ -20,6 +30,7 @@ window.config = {
       },
     ],
   },
+
   // Extensions should be able to suggest default values for these?
   // Or we can require that these be explicitly set
   hotkeys: [
@@ -124,6 +135,6 @@ window.config = {
   // If the server is particularly slow to respond to series metadata
   //  requests as it extracts the metadata from raw files everytime,
   //  try setting this to even lower value
-  // Leave it undefined for no limit, sutiable for HTTP/2 enabled servers
+  // Leave it undefined for no limit, suitable for HTTP/2 enabled servers
   // maxConcurrentMetadataRequests: 5,
 };